requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
python-dotenv
streamlit
chromadb
beautifulsoup4
transformers
langchain-groq
flask
fastapi
sentence-transformers
crewai
crewai_tools